Revert "Changed salary datatype from double to int"
This reverts commit 511e9daa6784af36a8ab167f16d3426ec80d2adc.

+++ src/com/java/Employee.java
@@ -8,7 +8,7 @@ public class Employee {
 	//Getters and Setters
 	private String name;
 	private int age;
-	private int salary;
+	private double salary;
 	private String address;
 	
 	public String getName() {
